chore: Updating MenuFlyout
